# relay-server/openapi.yaml
openapi: 3.1.0
info:
  title: Zenith Relay Compatible Server
  version: 1.0.0
  description: >
    Public management and Responses, Chat Completions, Messages and Gemini runtime contract for a server
    operated by the user. This server is separate from the production Zenith
    Gateway and Control API and has no contract for Zenith production secrets,
    customer keys, backend tokens, account-pool inventory, or internal
    production routing/business logic. Management snapshots, usage, source
    statistics, and diagnostics are redacted; raw credentials, cookies,
    authorization headers, prompts, and provider response bodies never appear
    in those management responses. The /v1 data plane may carry the user's
    request and provider response as required by the selected API contract.
servers:
  - url: https://relay.example.com
paths:
...
  /capabilities:
    get:
      operationId: capabilities
      security: [{ managementToken: [] }]
      responses:
        "200":
          description: Negotiated protocol versions and optional features
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capabilities"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /state:
    get:
      operationId: state
      security: [{ managementToken: [] }]
      responses:
        "200":
          description: >
            Canonical redacted runtime snapshot. It contains operational state
            and aggregates only; credentials, cookies, authorization headers,
            prompts, raw provider bodies, and vault material are excluded.
            gateway.modelCatalog optionally maps member model IDs to advisory
            catalogProvider and catalogFamily values, including excluded models.
            This display metadata does not grant runtime eligibility.
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
...
  securitySchemes:
    managementToken: { type: http, scheme: bearer, description: User-managed server administration token; never valid for /v1 and never a Zenith production credential. }
    profileCredential: { type: http, scheme: bearer, description: Relay-managed data-plane credential for /v1; never valid for management and never a Zenith production credential. }
    messagesCredential: { type: apiKey, in: header, name: x-api-key, description: Pool request key with Messages permission. }
    geminiCredential: { type: apiKey, in: header, name: x-goog-api-key, description: Pool request key with Gemini permission. }
  responses:
...
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { type: object }
  schemas:
...
    Capabilities:
      type: object
      required: [serverId, protocolMin, protocolMax, features]
      properties:
        serverId: { type: string, minLength: 1 }
        protocolMin: { type: integer, minimum: 1 }
        protocolMax: { type: integer, minimum: 1 }
        features: { type: array, items: { type: string }, uniqueItems: true, description: source_protocols_v1 enables protocolMode, endpoint evidence, computed routes and explicit generation checks. }
...